Issues/Pull Request: Hide banners
This change hides the dismissable banners that show on each repository's
issue and pull request pages which contain info only valuable for
newcomers.

While this might be a opinionated change, it is assumed that most
GitHub Dark users are power users who might also enjoy not having to
constantly click on these banners.