fix: retweak a8f3e69 re: #731
address the issue some areas hover that is not present in default.

The hover was coming from .highlight:not(.lines) tr:not(.inline-comments):hover td:not(.line-age)

Also had to add the excludes below to line above to cure part 1 of issue.
1. :not(.blob-num-hunk):not(.non-expandable)
2. added the .blob-code-hunk, .blob-code-expandable just below this new selector so the bg color 
   is same this cures part 2 of issue.

Im hoping the addition of these exclusions dont break anything
and this finally looks like default except these are grayscale.

tested in 1.20.63 by viewing commit 7e6a35b