Increase specificity for octotree, fixes #911
Currently, `.octotree-sidebar.octotree-github-sidebar .octotree-views .octotree-tree-view .jstree-anchor` has the same CSS specificity as the one from octotree itself, as both are defined using `!important` and with the same number of classes. Thus, which style is selected is not deterministic (GitHub-Dark is selected in Chrome, original Octotree is selected in Firefox).

I notice that all `jstree-anchor` are stored inside `jstree`, thus adding `.jstree` selector will increase specificity (since there are higher number of classes than the original Octotree) and be selected.