fix https://github.com/enterprise subheader see desc.
* fixes the mobile contact sales button styles
* fixes the hover states on subheader text

again this is a wack-a-mole can of worms with desktop vs mobile, so to ensure fix stays I used the @media queries which seems no one likes.

But there we are, its this or broken unless someone can come up with something better that actually works.

this was 3 weeks in the making, I tried various other ways without @media and none worked.