Fix pink border rules for sponsor alerts
The border of these elements wasn't being overridden since it was set with an inline style. Of the three spots I see these (sponsors dashboard alert, matching fund full width alert on your sponsors page, and the matching fund inline alert on your edit tiers page) the only spot not fully fixed is the edit tiers page because that inline style is using `!important`.

The text color was aso breaking the alerts on the sponsors pages and doesn't seem to be needed.