Hack to fix the feature preview button list color
These buttons don't set a background color so they default to `buttonface`. There aren't any useful classes on these buttons so using the one to style it on hover was the next best option.

We can't use the background rules above this one because they have `!important` which stops the selected & hover styles from working.