Add text color to generator hover style
This fixes the dropdown menu on https://developer.github.com/. The text color was being set to `#dcdcdc !important` from the `.site-header-nav a` rule. This adds the text color to the generator like we do for the other hover styles but it includes `!important` because of the other rule taking priority.