Add bg to notification comment loader container
On the new notification page there's a divider that shows when only some of an issue's comments are visible. The child elements have a background color and use top & bottom padding, but the second element uses padding & margin which results in a 4px gap where the background color shows through. This attempts to fix that without changing the layout.