Fix top border on pr checks view
The actions/checks views were recently updated so this issue only exists on the PR checks tab when no results are selected. If results are selected one of the child elements has a top border which is why we need the `:not()`. This can be seen at https://github.com/StylishThemes/GitHub-Dark/pull/959/checks